deleted empty line and rename interface getArea

+++ src/main/java/core/basesyntax/RightTriangle.java
@@ -6,8 +6,8 @@ public class RightTriangle extends Figure {
 
     public RightTriangle(Color color, int base, int height) {
         super(color);
-        this.base = (int) base;
-        this.height = (int) height;
+        this.base = base;
+        this.height = height;
     }
 
     @Override
@@ -18,7 +18,7 @@ public class RightTriangle extends Figure {
     @Override
     public void draw() {
         System.out.println("Figure: right triangle, area: " + getArea() + " sq.units, base: "
-                + (int) base + ", height: " + (int)height + " units, color: " + getColor());
+                + base + ", height: " + height + " units, color: " + getColor());
 
     }
 }
